+++ R/attach_source.R
@@ -1,4 +1,4 @@
-#' Source R files in an attached environment
+#' Source R files into an attached environment
 #'
 #' @param ... filepaths to R files, or paths to directories containing R files.
 #' @param name A string, the name for the attached environment. By default, the
@@ -10,46 +10,22 @@
 #'   one. If an environment of `name` already exists, `pos` is ignored.
 #' @param chdir logical. if TRUE, the \R working directory is temporarily
 #'   changed to the directory containing the file(s) being sourced.
+#' @param parent R environment. If an environment specified by `name` is not
+#'   already attached, then the supplied R scripts are first sourced into a new
+#'   environment with the supplied parent. The default of `globalenv()` enables
+#'   calling `library()` in the scripts and having the subsequent
+#'   code in the scripts "see" the attached packages.
 #' @param warn.conflicts logical. If TRUE (the default), print warnings about
 #'   objects in the attached environment that that are masking or masked by
-#'   other objects of the same name.
+#'   other objects of the same name. If the environment specified by `name` is
+#'   was attached previously, then only newly defined objects are warned about.
+#'   N.B., Even though the name is `warn.conflicts`, the messages about
+#'   conflicts are not `warning()`s but `packageStartupMessage()`s, and can be
+#'   suppressed with `suppressPackageStartupMessages()`
 #' @param mask.ok character vector of names of objects that can mask objects on
-#'   the search path without signaling a warning if `warn.conflicts` is `TRUE`
-#'
-#' @note One subtlety that is sometimes important: packages attached after this
-#'   environment is created will be not on the symbol search path for the environment
-#'   where the R source is evaluated. The search path of the environment the R
-#'   files are sourced in is `tail(search(), -pos)`.
-#'
-#'   This means that, for example, if you source a script that calls
-#'   `library()`, the code in that script will not "see" the functions from the
-#'   newly attached packages. This is by design. However, if you want to source
-#'   scripts that call `library` and define new functions, and you want those
-#'   new functions to "see" the `library` attached packages, here are 3 ways to
-#'   do that:
-#'
-#'   1.  Attach all the packages you want before attaching the script:
-#'       ````r
-#'       library(foo); library(bar)
-#'       attach_source("my_script.R")
-#'       ````
-#'
-#'   2.  Modify the default `pos` argument to `library`, so all new packages
-#'       attach after your script:
-#'       ````r
-#'       envir:::set_default_library_pos(after = "source:my_script.R")
-#'       attach_source("my_script.R")
-#'       ````
-#'
-#'   3.  This is the likely the most preferred solution. Instead of sourcing
-#'       files directly in the attached environment, source the files into a new
-#'       environment that inherits from `.Globalenv`, and then copy over everything
-#'       to the attached environment.
-#'       ````r
-#'       attach_eval({
-#'         import_from("my_script.R")
-#'       })
-#'       ````
+#'   the search path without signaling a warning if `warn.conflicts` is `TRUE`.
+#'   The sourced R script can also define `.mask.OK` in the R environment, which
+#'   has the same effect as passing it as an argument.
 #'
 #' @return The attached environment, invisibly.
 #' @seealso [import_from], [set_library_default_pos]
@@ -60,21 +36,69 @@ attach_source <- function(..., # files_andor_dirs,
                           pos = 2L,
                           chdir = FALSE,
                           warn.conflicts = TRUE,
-                          mask.ok = NULL) {
+                          mask.ok = NULL,
+                          parent = .GlobalEnv) {
 
-  if(!(is.character(name) && identical(length(name), 1L)))
+  if (!(is.character(name) && identical(length(name), 1L)))
     stop("`name` must be a string")
 
-  envir <- as_maybe_attached_env(name, pos)
 
-  if (warn.conflicts) {
-    mask.ok <- c(mask.ok, names(envir))
-    on.exit(warn_about_conflicts(envir, ignore = mask.ok))
+  if (name %in% search()) {
+    # env is already attached
+
+    attached_env <- as.environment(name)
+    mask.ok <- c(mask.ok, names(attached_env)) # don't warn about objects already masked
+    include(c(...), envir = attached_env, chdir = chdir, recursive = recursive)
+
+  } else if (is.null(parent)) {
+    # user is opt-ing out of new.env(parent=globalenv()), want "old-style" attach_source()
+    # where the script is literally sys.source()'d in the already attached env
+
+    attached_env <- attach(NULL,
+                           pos = pos,
+                           name = name,
+                           warn.conflicts = FALSE)
+    include(c(...), envir = attached_env, chdir = chdir, recursive = recursive)
+
+  } else {
+    # default under typical usage: create a new.env(), then attach it.
+    # (N.B: attach(env) doesn't actually modify env; it creates a
+    # new (attached) env then shallow copies all the bindings)
+
+    # This approach allows R scripts to contain library() calls and function
+    # definitions, and have the functions defined in the scripts "see" the
+    # library().
+    envir <- new.env(parent = parent)
+    attr(envir, "name") <- name # a visible reminder when printing functions
+    include(c(...), envir = envir, chdir = chdir, recursive = recursive)
+    attached_env <- attach(envir, pos = pos, name = name, warn.conflicts = FALSE)
   }
 
-  include(c(...), envir = envir, chdir = chdir, recursive = recursive)
-}
+  if (warn.conflicts) {
+    if (exists(".conflicts.OK", envir = attached_env, inherits = FALSE)) {
+      .conflicts.OK <- get0(".conflicts.OK", envir = attached_env, inherits = FALSE)
+      # attach() only checks for exists(".conflicts.OK"), and if TRUE, skips
+      # warnings alltogether. We allow for defining .conflicts.OK to a character
+      # vector of names to specifically not warn about -- essentially an alias
+      # for .mask.OK
+      if (is.character(.conflicts.OK) &&
+          length(.conflicts.OK) &&
+          length(intersect(.conflicts.OK, names(attached_env)))) {
+        mask.ok <- c(mask.ok, .conflicts.OK)
+      } else {
+        return(invisible(attached_env))
+      }
+    }
 
+    mask.ok <- c(mask.ok, get0(".mask.OK", envir = attached_env,
+                               inherits = FALSE))
+
+    warn_about_conflicts(attached_env, ignore = unique(mask.ok))
+  }
+
+
+  invisible(attached_env)
+}
 
 
 as_tidy_env_name <- function(x, prefix = NULL) {
@@ -101,7 +125,7 @@ warn_about_conflicts <- function(envir, ignore = NULL) {
               seq.int(from = envir_pos + 1, to = length(sp))))
     if (length(masked <- setdiff(intersect(names(pos.to.env(i)), objs),
                                  already_warned_about))) {
-      message(.maskedMsg(
+      packageStartupMessage(.maskedMsg(
         masked,
         pkg = sprintf("%s (pos = %i)", sp[i], i),
         by = i < envir_pos